[MODIFY] Command 기능 수정합니다.

+++ src/main/java/com/samsungds/codereview/teamd/command/DelCommand.java
@@ -8,7 +8,6 @@ import com.samsungds.codereview.teamd.vo.Employee;
 import java.io.IOException;
 import java.util.ArrayList;
 import java.util.Map;
-import java.util.TreeMap;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
@@ -27,32 +26,21 @@ public class DelCommand implements ICommand{
 
         if(itemList.size() != 6) throw new IllegalArgumentException("Error : Argument Count");
 
-        Map<Integer, Employee> map = irepo.delete(checkSearchKey(itemList.get(Constants.INPUT_STR_OPTION2_POS),
+        if(itemList.get(Constants.INPUT_STR_OPTION1_POS).equals(Constants.OPTION1_PRINT)){
+            Map<Integer, Employee> map = irepo.delete(checkSearchKey(itemList.get(Constants.INPUT_STR_OPTION2_POS),
+                    itemList.get(Constants.INPUT_STR_KEY1)), itemList.get(Constants.INPUT_STR_VALUE1));
+            printResult(transMaptoList(map));
+            return true;
+        }
+
+        int cnt = irepo.deleteCnt(checkSearchKey(itemList.get(Constants.INPUT_STR_OPTION2_POS),
                 itemList.get(Constants.INPUT_STR_KEY1)), itemList.get(Constants.INPUT_STR_VALUE1));
 
-        ArrayList<Employee> empList = new ArrayList<>();
-
-        // 임시 Sorting (Treemap 구조로 변경)
-        Map<Integer, Employee> map1 = new TreeMap<>();
-
-        if(map != null) {
-            for (Integer key : map.keySet()) {
-                map1.put(key, map.get(key));
-            }
-        }
-
-        // 임시 Limit 설정
-        int cnt = 0;
-        for (Integer key : map1.keySet()) {
-            empList.add(map1.get(key));
-            cnt++;
-            if(cnt == 5) break;
-        }
-
-        printResult(empList, isPrintOptionEnable(itemList.get(Constants.INPUT_STR_OPTION1_POS)));
-
+        printResult(cnt);
         return true;
     }
+
+
 
     @Override
     public void setFilePrint(Print filePrint){
@@ -91,11 +79,19 @@ public class DelCommand implements ICommand{
         return inputStrList;
     }
 
-    private void printResult(ArrayList<Employee> empList, Boolean isEnable) throws IOException {
-        filePrint.print(Constants.COMMAND_DEL, empList, isEnable);
+    private ArrayList<Employee> transMaptoList(Map<Integer, Employee> map) {
+        ArrayList<Employee> empList = new ArrayList<>();
+        for(Integer key : map.keySet()){
+            empList.add(map.get(key));
+        }
+        return empList;
     }
 
-    private Boolean isPrintOptionEnable(String inputStr){
-        return Constants.OPTION1_PRINT.equals(inputStr);
+    private void printResult(ArrayList<Employee> empList) throws IOException {
+        filePrint.print(Constants.COMMAND_DEL, empList);
+    }
+
+    private void printResult(int cnt) throws IOException {
+        filePrint.print(Constants.COMMAND_DEL, cnt);
     }
 }
